Updated Full UT Scraper!
I made the get_votes function and integrated the two bill scrapers, so it should work for any UT bill now. Ran into an issue with the metadata function for 2002 and added a line that might've fixed it. Also still having issues with the write file function for some reason, hoping you could help with that!